Removing googletest-release-1.7.0 third-party code from source code
Adding googletest release-1.7.0 as a git submodule. With this, managing third-
party source will be easier once if you want to test this code with a later
release from this library you should just checkout a new tag. The code of this
library if officially distributed by the submodule git repository:

https://github.com/google/googletest.git